Refactor: Reorganize ai-commands module structure
- Move CLI logic from mlflow/cli/ai_commands.py to mlflow/ai_commands/__init__.py
- Rename core.py to ai_command_utils.py for clarity
- Rename test files to match new module structure
- Update all imports to use new module paths
- Update package-data in pyproject.py generator
- Add comprehensive type hints throughout codebase
- Update help text to consistently say 'AI commands'
