+++ libs/skinny/pyproject.toml
@@ -128,6 +128,7 @@ mlflow = [
   "server/auth/basic_auth.ini",
   "server/auth/db/migrations/alembic.ini",
   "models/notebook_resources/**/*",
+  "ai_commands/**/*.md",
 ]
 
 [tool.setuptools.packages.find]
