+++ mlflow/ai_commands/__init__.py
@@ -1,26 +1,28 @@
-"""CLI commands for managing MLflow commands."""
+"""CLI commands for managing MLflow AI commands."""
 
 import click
 
-from mlflow.commands import get_command, list_commands
+from mlflow.ai_commands.ai_command_utils import get_command, list_commands, parse_frontmatter
+
+__all__ = ["get_command", "list_commands", "parse_frontmatter", "commands"]
 
 
 @click.group("ai-commands")
-def commands():
+def commands() -> None:
     """Manage MLflow AI commands for LLMs."""
 
 
 @commands.command("list")
 @click.option("--namespace", help="Filter commands by namespace")
-def list_cmd(namespace):
-    """List all available commands."""
+def list_cmd(namespace: str | None) -> None:
+    """List all available AI commands."""
     cmd_list = list_commands(namespace)
 
     if not cmd_list:
         if namespace:
-            click.echo(f"No commands found in namespace '{namespace}'")
+            click.echo(f"No AI commands found in namespace '{namespace}'")
         else:
-            click.echo("No commands found")
+            click.echo("No AI commands found")
         return
 
     for cmd in cmd_list:
@@ -29,8 +31,8 @@ def list_cmd(namespace):
 
 @commands.command("get")
 @click.argument("key")
-def get_cmd(key):
-    """Get a specific command by key."""
+def get_cmd(key: str) -> None:
+    """Get a specific AI command by key."""
     try:
         content = get_command(key)
         click.echo(content)
