+++ pyproject.release.toml
@@ -129,6 +129,7 @@ mlflow = [
   "server/auth/basic_auth.ini",
   "server/auth/db/migrations/alembic.ini",
   "models/notebook_resources/**/*",
+  "ai_commands/**/*.md",
   "models/container/**/*",
   "server/js/build/**/*",
 ]
