+++ pyproject.toml
@@ -145,6 +145,7 @@ mlflow = [
   "server/auth/basic_auth.ini",
   "server/auth/db/migrations/alembic.ini",
   "models/notebook_resources/**/*",
+  "ai_commands/**/*.md",
   "models/container/**/*",
   "server/js/build/**/*",
 ]
